Move tests to test directory (to avoid re-use of history object)
